Created TableofContents component
